refactor(stella-cli): the claude worker gets its own module, and its refusal a test
`work.rs` crossed the 1500-line ceiling, which is closed to growth and takes
no baseline entry. Everything specific to claude moves to
`self_driving_cmd/claude_worker.rs`: how its command line is built, how its
output is read, and the one thing it cannot honour. The parent keeps the
choice of worker and still measures the outcome from the tree.

A flat sibling rather than `work/claude.rs`, because `module-layout` blocks a
new `foo.rs`-beside-`foo/` pair — the rule AGENTS.md's Code style section now
states the opposite of. The crate's own layout is flat siblings, so this is
the neighborhood either way; the guard's disagreement with the prose it cites
is filed rather than fixed here.

The spend-limit refusal becomes `uncappable`, so what an operator is told when
a dollar cap meets a claude worker is asserted rather than assumed. It was the
one behaviour this change ships that no test could reach.
